+++ ad-tag/source/ts/ads/a9.ts
@@ -19,6 +19,9 @@ import { apstag } from '../types/apstag';
 import { tcfapi } from '../types/tcfapi';
 import TCPurpose = tcfapi.responses.TCPurpose;
 import * as adUnitPath from './adUnitPath';
+import prebid from 'prebid.js';
+import { prebidjs } from '../types/prebidjs';
+import IPrebidJs = prebidjs.IPrebidJs;
 
 const isA9SlotDefinition = (
   slotDefinition: Moli.SlotDefinition
@@ -60,6 +63,9 @@ export const a9Init = (
     'a9-init',
     (context: AdPipelineContext) =>
       new Promise<void>(resolve => {
+        // We need to report prebid here since it provides the convertCurrency function, we use for the floorprices.
+        context.window.pbjs = context.window.pbjs || ({ que: [] } as unknown as IPrebidJs);
+
         context.window.apstag = context.window.apstag || {
           _Q: [],
           init: function (): void {
@@ -238,7 +244,15 @@ export const a9RequestBids = (config: Moli.headerbidding.A9Config): RequestBidsS
                         // THe floor price is sent in EUR, amazon requires Cents
                         {
                           floor: {
-                            value: Math.ceil(priceRule.floorprice * 100 * 1.19),
+                            value: Math.ceil(
+                              context.window.pbjs?.convertCurrency
+                                ? context.window.pbjs.convertCurrency(
+                                    priceRule.floorprice,
+                                    'EUR',
+                                    'USD'
+                                  ) * 100
+                                : priceRule.floorprice * 100 * 1.19
+                            ),
                             currency: config.floorPriceCurrency || 'USD'
                           }
                         }